Implemented saving and opening of documents.
This required a rethink of the model paradigm. The model remains an array of DrawObjects. Previously, the DrawingView maintained this array. Now the document maintains the DrawObjects array.
Previously, when the app attempted to open a document, it would unarchive [DrawObject] and attempt to assign it to drawingView.drawObjects only to find that at the time of unarchiving the array, the nib had not been loaded and drawingView == nil.
By having the document maintain the model object, we avoid this problem.

Implemented alpha choice in the color panel.

Resolved bug where user could choose a color outside the RGBA colorspace, leading to an exception and no drawing of the proposed object while the user is drawing it.

DrawDocument.swift:

+ (line 25)				added drawingView outlet
+ (line 29)				added drawObjects model array
+ (lines 41 - 45)			added line to allow alpha choice in the color panel
+ (line 57)				implemented functioning dataOfType(_:error:) method
+ (lines 61 - 63)			implemented functioning readFromData(_:ofType:error:) method

DrawDocument.swift:

+ (line 9)				connected drawingView outlet

DrawingView.swift:

+ (lines 27 - 28)			removed model object drawObjects from the view
+ (line 48)				replaced drawObjects with document.drawObjects
+ (lines 111 - 112)		replaced repetitive code with call to helper method
+ (line 171)			eplaced drawObjects with document.drawObjects
+ (lines 185 - 219)		implemented helper method proposedColorForColor(_:), resolving bug where user could choose non RGBA color and generate exception

DrawObject.swift:

+ (line 22)				declared DrawObject as conforming to NSCoding
+ (lines 32 - 35)			implemented init(coder aDecoder: NSCoder)
+ (lines 37 - 40)			implemented encodeWithCoder(_:)

Oval.swift:

+ (lines 28 - 31)			implemented init(coder aDecoder: NSCoder)
+ (lines 33 - 36)			implemented encodeWithCoder(_:)

Rectangle.swift:

+ (lines 28 - 31)			implemented init(coder aDecoder: NSCoder)
+ (lines 33 - 36)			implemented encodeWithCoder(_:)

Line.swift:

+ (lines 30 - 34)			implemented init(coder aDecoder: NSCoder)
+ (lines 36 - 40)			implemented encodeWithCoder(_:)